requirements.txt
beautifulsoup4
bitmath
colorama
configparser
lxml
pushbullet.py
termcolor
unicodecsv
